updated test cases
